Matrix multiplication

+++ model/Vector.java
@@ -1,17 +1,17 @@
 package model;
 
 public class Vector {
-    private int x;
-    private int y;
-    private int z;
+    private double x;
+    private double y;
+    private double z;
 
-    public Vector(int x, int y, int z) {
+    public Vector(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public void setVector(int x, int y, int z) {
+    public void setVector(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -21,15 +21,15 @@ public class Vector {
         return this;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 }
